Add quiz unittest
