+++ app/templates/quiz.html
@@ -5,7 +5,7 @@
 {% block content %}
 <div class="quiz-container">
     <div id="quiz" class="justify-center flex-column">
-        <h1>QUIZ TITLE</h1>
+        <h1>TAILORED QUIZ</h1>
         <h2 id="question">{{question}}</h2>
 
         {% if edu %}
